Antoine : Implémentation methodes produits dao et service

+++ 02_GitProjet/src/fr/adaming/entities/Produit.java
@@ -36,7 +36,7 @@ public class Produit {
 	}
 
 	// Getters et setters
-	public long getIdProduit() {
+	public int getIdProduit() {
 		return idProduit;
 	}
